Update build process and dependencies
Update the devDependencies - update React and Jest libraries.
With the update to Jest, they deprecated some config options in package.json
suck as scriptProcessor. Remove testIgnorPattern and scriptPreprocessor
and add moduleNameMapper and unmockedModulePatterns. Add __mocks__
directory and styleMock.js to be able to mock the scss
that gets imported in to the component to ignore the styles.

Change dev to deveopment and prod to production in
package.json and webpack configs. Add __DEV__ variable
to eslint globals and Jest globals