Refactor component to be stateless
Refactor the Tags component to be stateless and allow the state of
tags to be changed/controlled at the host component level. Keep all
other business logic in the component layer.

Update README to reflect new refactor of the component. Update
options and events, remove onChange event.

Update tags tests to use stateless compoment. Create an
App fixture to be able to capture events and state change
in the component so we can test properly.

Update webpack to extract the css in a production build
as well as a development build. Remove css-loader rule from rules config.

Update example/docs to reflect new changes to the component.